+++ fe/fe-core/src/main/java/com/starrocks/statistic/columns/ColumnUsage.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.starrocks.statistic.predicate_columns;
+package com.starrocks.statistic.columns;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Splitter;
