+++ fe/fe-core/src/main/java/com/starrocks/statistic/columns/PredicateColumnsMgr.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.starrocks.statistic.predicate_columns;
+package com.starrocks.statistic.columns;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Maps;
