+++ fe/fe-core/src/main/java/com/starrocks/statistic/columns/PredicateColumnsStorage.java
@@ -1,4 +1,5 @@
-// Copyright 2021-present StarRocks, Inc. All rights reserved. //
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.starrocks.statistic.predicate_columns;
+package com.starrocks.statistic.columns;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Stopwatch;
